# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Heave Games API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: authLogin
      tags: [auth]
      summary: Login with username/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Unauthorized
...
  /api-keys:
    get:
      operationId: listApiKeys
      tags: [api-keys]
      summary: List all API keys
      responses:
        "200":
          description: List of API keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiKey"
    post:
      operationId: createApiKey
      tags: [api-keys]
      summary: Create API key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiKeyInput"
      responses:
        "201":
          description: Created API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKey"
  /api-keys/{id}:
    patch:
      operationId: updateApiKey
      tags: [api-keys]
      summary: Update API key
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiKeyUpdate"
      responses:
        "200":
          description: Updated API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKey"
    delete:
      operationId: deleteApiKey
      tags: [api-keys]
      summary: Delete API key
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /api-keys/{id}/revoke:
    post:
      operationId: revokeApiKey
      tags: [api-keys]
      summary: Revoke API key
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKey"

  # CATEGORIES
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List categories
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ENDPOINTS (API endpoint management)
  /endpoints:
    get:
      operationId: listApiEndpoints
      tags: [endpoints]
      summary: List all API endpoints
      responses:
        "200":
          description: Endpoints
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiEndpoint"
    post:
      operationId: createApiEndpoint
      tags: [endpoints]
      summary: Create a new endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiEndpointInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEndpoint"
  /endpoints/{id}:
    patch:
      operationId: updateApiEndpoint
      tags: [endpoints]
      summary: Update endpoint
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiEndpointUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEndpoint"
    delete:
      operationId: deleteApiEndpoint
      tags: [endpoints]
      summary: Delete endpoint
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /endpoints/{id}/toggle:
    post:
      operationId: toggleApiEndpoint
      tags: [endpoints]
      summary: Toggle endpoint active/inactive
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEndpoint"

  # IMAGES
...
  /backups/{id}/restore:
    post:
      operationId: restoreBackup
      tags: [backups]
      summary: Restore a backup
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Restored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Backup"

  # CONFIG
  /config:
    get:
      operationId: getConfig
      tags: [config]
      summary: Get global config
      responses:
        "200":
          description: Config
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConfigEntry"
    post:
      operationId: upsertConfig
      tags: [config]
      summary: Create or update a config entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfigInput"
      responses:
        "200":
          description: Updated config entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigEntry"
  /config/{key}:
    delete:
      operationId: deleteConfig
      tags: [config]
      summary: Delete config entry
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # GIVEAWAYS / SORTEOS
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
...
    AuthResult:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    # USER
    User:
      type: object
      required: [id, username, email, role, active, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [owner, admin, moderator, user]
        active:
          type: boolean
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
...
    ApiKey:
      type: object
      required: [id, name, keyPreview, active, createdAt, requestCount]
      properties:
        id:
          type: integer
        name:
          type: string
        keyPreview:
          type: string
        active:
          type: boolean
        requestCount:
          type: integer
        rateLimit:
          type: ["integer", "null"]
        expiresAt:
          type: ["string", "null"]
        createdAt:
          type: string
        userId:
          type: ["integer", "null"]
    ApiKeyInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        rateLimit:
          type: ["integer", "null"]
        expiresAt:
          type: ["string", "null"]
        userId:
          type: ["integer", "null"]
    ApiKeyUpdate:
      type: object
      properties:
        name:
          type: string
        active:
          type: boolean
        rateLimit:
          type: ["integer", "null"]

    # CATEGORY
    Category:
      type: object
      required: [id, name, slug, description, icon, active, itemCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        icon:
          type: string
        active:
          type: boolean
        itemCount:
          type: integer
        createdAt:
          type: string
    CategoryInput:
      type: object
      required: [name, slug, description, icon]
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        icon:
          type: string
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        active:
          type: boolean

    # API ENDPOINT
    ApiEndpoint:
      type: object
      required: [id, path, method, description, active, categoryId, requestCount, responseJson, createdAt]
      properties:
        id:
          type: integer
        path:
          type: string
        method:
          type: string
        description:
          type: string
        active:
          type: boolean
        categoryId:
          type: integer
        requestCount:
          type: integer
        responseJson:
          type: string
        requiresAuth:
          type: boolean
        rateLimit:
          type: ["integer", "null"]
        createdAt:
          type: string
    ApiEndpointInput:
      type: object
      required: [path, method, description, categoryId, responseJson]
      properties:
        path:
          type: string
        method:
          type: string
          enum: [GET, POST, PUT, PATCH, DELETE]
        description:
          type: string
        categoryId:
          type: integer
        responseJson:
          type: string
        requiresAuth:
          type: boolean
        rateLimit:
          type: ["integer", "null"]
    ApiEndpointUpdate:
      type: object
      properties:
        description:
          type: string
        active:
          type: boolean
        responseJson:
          type: string
        requiresAuth:
          type: boolean
        rateLimit:
          type: ["integer", "null"]

    # IMAGE
...
    Backup:
      type: object
      required: [id, name, size, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        size:
          type: string
        status:
          type: string
          enum: [completed, in_progress, failed]
        createdAt:
          type: string
        restoredAt:
          type: ["string", "null"]

    # CONFIG
    ConfigEntry:
      type: object
      required: [id, key, value, type, description, createdAt]
      properties:
        id:
          type: integer
        key:
          type: string
        value:
          type: string
        type:
          type: string
          enum: [string, number, boolean, json]
        description:
          type: string
        isSecret:
          type: boolean
        createdAt:
          type: string
    ConfigInput:
      type: object
      required: [key, value, type]
      properties:
        key:
          type: string
        value:
          type: string
        type:
          type: string
          enum: [string, number, boolean, json]
        description:
          type: string
        isSecret:
          type: boolean

    # GIVEAWAY
...