# lib/api-zod/src/generated/types/apiEndpoint.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Heave Games API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ApiEndpoint {
  id: number;
  path: string;
  method: string;
  description: string;
  active: boolean;
  categoryId: number;
  requestCount: number;
  responseJson: string;
  requiresAuth?: boolean;
  /** @nullable */
  rateLimit?: number | null;
  createdAt: string;
}