# lib/api-zod/src/generated/types/apiEndpointInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Heave Games API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ApiEndpointInputMethod } from './apiEndpointInputMethod';

export interface ApiEndpointInput {
  path: string;
  method: ApiEndpointInputMethod;
  description: string;
  categoryId: number;
  responseJson: string;
  requiresAuth?: boolean;
  /** @nullable */
  rateLimit?: number | null;
}