# lib/api-zod/src/generated/types/apiEndpointUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Heave Games API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ApiEndpointUpdate {
  description?: string;
  active?: boolean;
  responseJson?: string;
  requiresAuth?: boolean;
  /** @nullable */
  rateLimit?: number | null;
}